+++ README.md
@@ -1,22 +1,22 @@
 # mod_auth_tkt cookie implementation
 
-[![Build Status](https://travis-ci.org/yola/authtkt.png)](https://travis-ci.org/yola/authtkt)
+[![Build Status](https://travis-ci.org/yola/auth_tkt.png)](https://travis-ci.org/yola/auth_tkt)
 
 ## Modules
 
-`authtkt.ticket` is a Python re-implementation of the [mod_auth_tkt][]
+`auth_tkt.ticket` is a Python re-implementation of the [mod_auth_tkt][]
 cookie. Cookies can be created with `AuthTkt` and verified with
 `verify`.
 
 [mod_auth_tkt]: http://www.openfusion.com.au/labs/mod_auth_tkt/
 
-`authtkt.encrypted`'s `EncryptedAuthTkt` is a wrapper around `AuthTkt`
+`auth_tkt.encrypted`'s `EncryptedAuthTkt` is a wrapper around `AuthTkt`
 that stores an encrypted JSON payload in the mod_auth_tkt cookie's
 user-data section.
 
 ## Helpers
 
-There is a `get_ticket_data` function in `authtkt.helepers`, that
+There is a `get_ticket_data` function in `auth_tkt.helepers`, that
 decrypts and verifies a cookie.
 
 ## Testing
